Add UK government API catalogue
This isn't an API but rather the site where they are collated by the UK
government.

I've added the data.gov.uk site as a separate entry for clarity (because
of the number of similar sites from other countries). This site is also
linked to from the API catalogue site.